requirements.txt
streamlit
pandas
numpy
scikit-learn
catboost
shap
matplotlib
qrcode
pillow